[packages/template,core][s]: move Nav
